+++ src/Models/Concerns/HasIcon.php
@@ -3,6 +3,7 @@
 namespace Waterhole\Models\Concerns;
 
 use Illuminate\Http\UploadedFile;
+use Intervention\Image\Encoders\PngEncoder;
 use Intervention\Image\Image;
 use Waterhole\Models\Attributes\FileAttribute;
 
@@ -55,7 +56,7 @@ trait HasIcon
         return $this->fileAttribute(
             attribute: 'icon_file',
             directory: 'icons',
-            encodeImage: fn(Image $image) => $image->scaleDown(50, 50)->toPng(),
+            encodeImage: fn(Image $image) => $image->scaleDown(50, 50)->encode(new PngEncoder()),
         );
     }